Fixed issues with Lotto Fixed issue with Lotto time Fixed issue with Bank value saving Fixed Pay not adding to Bank Updated Lotto to deposit into bank Updated stats to be ulong

+++ Credify/Commands/BuyCommand.cs
@@ -88,7 +88,7 @@ public class BuyCommand : Command
             clientItem.Amount++;
         }
         
-        _persistenceManager.StatisticsState.CreditsSpent += serverItem.Cost;
+        _persistenceManager.StatisticsState.CreditsSpent += (ulong)serverItem.Cost;
         await _persistenceManager.AlterClientCredits(-serverItem.Cost, client: gameEvent.Origin);
         await _persistenceManager.WriteClientShopAsync(gameEvent.Origin, clientItems);
         gameEvent.Origin.Tell(_credifyConfig.Translations.BoughtItem
